separate PageBulkExportJobStreamManager to a different file
