+++ apps/app/src/features/page-bulk-export/server/service/page-bulk-export/errors.ts
@@ -1,0 +1,15 @@
+export class DuplicateBulkExportJobError extends Error {
+
+  constructor() {
+    super('Duplicate bulk export job is in progress');
+  }
+
+}
+
+export class BulkExportJobExpiredError extends Error {
+
+  constructor() {
+    super('Bulk export job has expired');
+  }
+
+}
